Exclude DtStart if not in rrule
From the RFC 5545:
> The "DTSTART" property value SHOULD match the pattern of the
> recurrence rule, if specified. The recurrence set generated
> with a "DTSTART" property value that doesn't match the
> pattern of the rule is undefined.

https://datatracker.ietf.org/doc/html/rfc5545#page-120

Given the start date SHOULD match the pattern, it makes more
sense to ignore it if the value does not match the pattern.

A reason for the start date not matching is that it is not
known or easily determined until after generating occurrences
for the rule.