Evaluate calendar in a time zone always
This evaluates events in the calendar for a specific time zone.
Events with a time zone are evaluated in that event's time zone and then
converted to the requested time zone. Floating events are evaluated
in the request time zone. This is essentially the same behavior as before,
except the results are actual point-in-time values.

A key evaluation difference is that during backward daylight saving,
the last time offset can be output instead of only the first. Before,
an hourly repeating event would skip the second hour.