Change tests to fit zoned API
The few tests still failing were left alone on purpose. With some tests,
the test data needs to be changed because it does not make sense now that
all occurrences are in a specific time zone.

For FREEBUSY tests, that was mostly ignored because I am not sure how it
should be used.